arqcp a correr

+++ ARQCP/Total/arqcp.h
@@ -26,7 +26,7 @@ int limite_sensor(int maximo, int minimo, int valor);
 char escolher_max_min(float max, float min, float valor, char tentativas, Sensor *sensores);
 char escolher_frequencia(unsigned long freq2);
 void show_stats(float matrix[6][3], int rows, int cols);
-void matriz_diaria(int periodo, int * sensor, float matriz[6][3], int linha, int limitesuperior, int limiteinferior);
+void matriz_diaria(int periodo,unsigned short * sensor, float matriz[6][3], int linha, int limitesuperior, int limiteinferior);
 void exportCsv (Sensor* array, int size, float matrix[6][3]);
 
 #endif
